instantiate a new store for every component instance
