+++ src/lib/components/SectionListItem.svelte
@@ -1,8 +1,6 @@
 <script lang="ts">
-	import { goto } from '$app/navigation';
 	import { page } from '$app/stores';
 	import { Sitemap } from '$lib/sitemap';
-	import { sessionsStore, deleteStoreItem, knowledgeStore } from '$lib/store';
 	import { writable } from 'svelte/store';
 	import ButtonDelete from './ButtonDelete.svelte';
 	import Metadata from './Metadata.svelte';
@@ -11,27 +9,9 @@
 	export let id: string;
 	export let title: string;
 	export let subtitle: string;
+
 	const isSession = sitemap === Sitemap.SESSIONS;
-
-	// Create a unique store for this instance
 	const shouldConfirmDeletion = writable(false);
-
-	function deleteRecord() {
-		$shouldConfirmDeletion = false;
-
-		switch (sitemap) {
-			case Sitemap.KNOWLEDGE:
-				if ($knowledgeStore) $knowledgeStore = deleteStoreItem($knowledgeStore, id);
-				return goto('/knowledge');
-
-			case Sitemap.SESSIONS:
-				if ($sessionsStore) $sessionsStore = deleteStoreItem($sessionsStore, id);
-				return goto('/sessions');
-
-			default:
-				break;
-		}
-	}
 </script>
 
 <!-- Need to use `#key id` to re-render the delete nav after deletion -->
@@ -58,7 +38,7 @@
 			class="section-list-item__delete"
 			class:section-list-item__delete--confirm-deletion={$shouldConfirmDeletion}
 		>
-			<ButtonDelete {deleteRecord} {shouldConfirmDeletion} />
+			<ButtonDelete {sitemap} {id} {shouldConfirmDeletion} />
 		</nav>
 	</div>
 {/key}
